+++ src/main/java/kono/ene/napi/dao/repository/NintendoAuthDao.java
@@ -5,5 +5,5 @@ import kono.ene.napi.dao.base.BaseRepo;
 import kono.ene.napi.dao.entity.AuthDo;
 
 public interface NintendoAuthDao extends BaseRepo<AuthDo, String> {
-    AuthDo findByQid(Integer qid);
+    AuthDo findByQid(Long qid);
 }
