+++ src/main/java/kono/ene/napi/dao/repository/NintendoSwitchUserDao.java
@@ -5,5 +5,5 @@ import kono.ene.napi.dao.base.BaseRepo;
 import kono.ene.napi.dao.entity.SwitchUserDo;
 
 public interface NintendoSwitchUserDao extends BaseRepo<SwitchUserDo, String> {
-    SwitchUserDo findByQid(Integer qid);
+    SwitchUserDo findByQid(Long qid);
 }
