+++ src/main/java/kono/ene/napi/dao/repository/NintendoSwitchWebAccessTokenDao.java
@@ -5,5 +5,5 @@ import kono.ene.napi.dao.base.BaseRepo;
 import kono.ene.napi.dao.entity.WebAccessTokenDo;
 
 public interface NintendoSwitchWebAccessTokenDao extends BaseRepo<WebAccessTokenDo, String> {
-    WebAccessTokenDo findByQidAndGameId(Integer qid, Long gameId);
+    WebAccessTokenDo findByQidAndGameId(Long qid, Long gameId);
 }
