+++ src/main/java/kono/ene/napi/dao/repository/NintendoUserDao.java
@@ -5,5 +5,5 @@ import kono.ene.napi.dao.base.BaseRepo;
 import kono.ene.napi.dao.entity.UserDo;
 
 public interface NintendoUserDao extends BaseRepo<UserDo, String> {
-    UserDo findByQid(Integer qid);
+    UserDo findByQid(Long qid);
 }
