+++ src/main/java/kono/ene/napi/request/AccountAccessTokenRequest.java
@@ -8,5 +8,5 @@ import lombok.NoArgsConstructor;
 @AllArgsConstructor
 @NoArgsConstructor
 public class AccountAccessTokenRequest {
-    private Integer qid;
+    private Long qid;
 }
