+++ src/main/java/kono/ene/napi/request/UserInfoRequest.java
@@ -8,5 +8,5 @@ import lombok.NoArgsConstructor;
 @NoArgsConstructor
 @AllArgsConstructor
 public class UserInfoRequest {
-    private Integer qid;
+    private Long qid;
 }
